First attempt at networking
I have so far attempted implement the basics of sending files over a
network using FastAPI. I haven't setup auto-discovery, so an IP and port
must be passed for this to work (again untested at this point). I then
need to add delta syncing, as in, only sync the changes and not the
whole file.